+++ root/uicommon.py
@@ -665,7 +665,7 @@ class PriceInputDialog(ui.ScriptWindow):
 		money = 0
 		if text and text.isdigit():
 			try:
-				money = long(text)
+				money = int(text)
 			except ValueError:
 				money = 199999999
 		self.moneyText.SetText(localeinfo.NumberToMoneyString(money))
