+++ root/uipickmoney.py
@@ -81,7 +81,7 @@ class PickMoneyDialog(ui.ScriptWindow):
 		text = self.pickValueEditLine.GetText()
 
 		if len(text) > 0 and text.isdigit():
-			money = long(text)
+			money = int(text)
 			money = min(money, self.maxValue)
 
 			if money > 0:
